+++ index.html
@@ -109,7 +109,6 @@
     <footer class="copy">
         <p>Todos os direitos reservados</p>
     </footer>
-<script src="./animacoes.js"></script>
-<script src="./script.js"></script>
+<script type="module" src="./js/script.js"></script>
 </body>
 </html>
