+++ js/modules/initTabNav.js
@@ -1,0 +1,19 @@
+export default function initTabNav() {
+    const tabMenu = document.querySelectorAll('.js-tabmenu li');
+    const tabContent = document.querySelectorAll('.js-tabcontent section');
+    const activeClass = 'ativo';
+    if(tabMenu.length && tabContent.length){
+        tabContent[0].classList.add(activeClass);
+        function activeTab (index){
+            tabContent.forEach((section) => {
+                section.classList.remove(activeClass);
+            });
+            tabContent[index].classList.add(activeClass);
+        }
+        tabMenu.forEach((itemMenu, index) =>{
+            itemMenu.addEventListener('click', () =>{
+                activeTab(index);
+            })
+        });
+    }
+}
