+++ js/script.js
@@ -1,0 +1,9 @@
+import initTabNav from "./modules/initTabNav.js";
+import initAccordion from "./modules/initAccordion.js";
+import smoothScroll from "./modules/smoothScroll.js";
+import initAnimacaoScroll from "./modules/initAnimacaoScroll.js";
+
+initTabNav();
+initAccordion();
+smoothScroll();
+initAnimacaoScroll();
